+++ modules/alloy/src/main/ts/ephox/alloy/api/ui/SplitFloatingToolbar.ts
@@ -50,7 +50,8 @@ const factory: CompositeSketchFactory<SplitFloatingToolbarDetail, SplitFloatingT
       parts: {
         button: externals['overflow-button'](),
         toolbar: externals.overflow()
-      }
+      },
+      onToggled: detail.onToggled,
     })
   );